ssl_ciphersuites: move internal functions declarations to a private header
