+++ library/ssl_misc.h
@@ -44,6 +44,7 @@
 #endif
 
 #include "mbedtls/pk.h"
+#include "ssl_ciphersuites_internal.h"
 #include "pk_internal.h"
 #include "common.h"
